requirements.txt
pandas==1.3.3
nltk==3.7
textblob==0.17.1
requests
sklearn
pyspellchecker==0.7.0
autocorrect==2.6.1
openpyxl==3.0.10